remove murmurhash; it's currently unused
It would be interesting to try out hash tables at some point, but it
introduces some complications. Namely, there are huge numbers of approaches
to hash tables (chaining, linear / quadratic probing, subtypes such as
robin hood hashing, perfect hashes when input set is known up front) and
many tunables to consider:

 * How many buckets to allocate initially?
 * Will the table grow?
 * What is the growth factor?
 * What hash algorithm should we use?

These questions indicate a large number of potential configurations. I would
like to investigate this, but getting results that can be plotted is a higher
priority right now. (Not that this repo is a priority.)